Added fixed grid
Debugged a problem with the grid not being passed as an argument to the check position function which arose after wrapping the main code into the "main()" function. Added fixed limits to show the whole grid in the graph every time. Corrected the y offset to make the graph the same as the 2d array.